fix on Dept

+++ LeaveApplicationTeam5/src/main/java/sg/edu/LeaveApplication/controller/DepartmentController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import sg.edu.LeaveApplication.model.Department;
 import sg.edu.LeaveApplication.service.DepartmentService;
 import sg.edu.LeaveApplication.service.DepartmentServiceImpl;
+import sg.edu.LeaveApplication.service.UserService;
 
 
 
@@ -28,6 +29,12 @@ public class DepartmentController {
 	{
 		this.dservice = dserviceImpl;
 	}
+	@Autowired
+	UserService uservice;
+	@Autowired
+	public void setUservice(UserService uservice) {
+		this.uservice = uservice;
+	}
 
 	
 	@RequestMapping (value = "/dplist" )
@@ -36,50 +43,52 @@ public class DepartmentController {
 		System.out.println(dservice.findAll().size());
 		model.addAttribute("dlist" , dservice.findAll());	
 		
-		return "/admin/department-list";
+		return "/admin/departmentList";
 			
 		}
-	
+
+
 	@RequestMapping (value = "/adddp" )
 		public String addForm(Model model)
 		{
 		model.addAttribute("department" ,new Department());	
-		return "/admin/department-form";
+		return "/admin/createDepartment";
 		}
 		
 	@RequestMapping (value = "/editdp/{id}" )
 		public String editForm(@PathVariable("id") Integer id, Model model)
 		
 		{
-		model.addAttribute("department", dservice.findDeparmentById(id));	
-		return "/admin/department-form";
+			model.addAttribute("department", dservice.findDeparmentById(id));	
+			return "/admin/createDepartment";
 		}
 		
 	@RequestMapping (value = "/savedp" )
 		public String saveDepartment(@ModelAttribute ("department") 
-		@Valid Department department, BindingResult bindingResult, Model model)
-		// @valid is done to make sure there is a valid department, and must
-		// do the binding result immediately after the object u wanna bind
-	
+		@Valid Department department, BindingResult bindingResult, Model model)	
 		{
 		if (bindingResult.hasErrors())
 		{
-			return "/admin/department-form";
+			return "/admin/createDepartment";
 		}
 		else
 		{
 			dservice.saveDepartment(department);
 		
-			return "forward:/dept/list";
+			return "redirect:/admin/dplist";
 		}
 		}
 		
 	@RequestMapping (value = "/deletedp/{id}" )
-		public String deleteDepartment(@PathVariable ("id") Integer id)
+		public String deleteDepartment(@PathVariable ("id") Integer id, Model model)
 		{
-		dservice.deleteDepartment(dservice.findDeparmentById(id));
-		
-		return "forward:/dept/list";
+			if(uservice.findUserBydepartment(dservice.findDeparmentById(id)).isEmpty()) {
+				dservice.deleteDepartment(dservice.findDeparmentById(id));
+			}
+			else {
+				model.addAttribute("msg", "Cannot delete department, please remove the users first.");
+			}
+		return "redirect:/admin/dplist";
 		}
